+++ PIGGISWS/Models/Rutero.cs
@@ -4,7 +4,7 @@ public class Rutero
 {
     public int RUT_EMPRESA {  get; set; }
     public DateTime RUT_FECHA { get; set; }
-    public int RUT_AGENTE { get; set; }
+    public decimal RUT_AGENTE { get; set; }
     public decimal RUT_CLIENTE { get; set; }
     public int RUT_VISITA { get; set; }
     public int RUT_PEDIDO { get; set; }
